Refactor Github Action generation
