+++ .github/workflows/develop_install.yml
@@ -30,8 +30,6 @@ jobs:
             conda_python_version: '3.8'
           - pytorch_version: '1.3.1'
             conda_python_version: '3.8'
-          - pytorch_version: '1.1.0'
-            jit_status: 'jit_disabled'
 
 
     steps:
